# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hamle Elementary School Student Information System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: getStudents
      tags: [students]
      summary: Get students (filtered by role)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a single student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student (admin/teacher)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentRequest"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Student deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Address:
      type: object
      properties:
        region:
          type: string
        zone:
          type: string
        kebele:
          type: string
        houseNo:
          type: string
      required:
        - region
        - zone
        - kebele
        - houseNo
    Student:
      type: object
      properties:
        id:
          type: string
        studentNo:
          type: string
        firstName:
          type: string
          pattern: '^[a-zA-Z\s]+$'
        lastName:
          type: string
          pattern: '^[a-zA-Z\s]+$'
        grade:
          type: string
        dateOfBirth:
          type: string
          pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
        gender:
          type: string
          enum: [male, female]
        address:
          $ref: '#/components/schemas/Address'
          nullable: true
        faydaId:
          type: string
          pattern: '^[0-9]{12}$'
          nullable: true
        enrollmentStatus:
          type: string
          enum: [active, inactive, withdrawn, transferred, graduated]
        medicalInfo:
          type: string
          nullable: true
        emergencyContact:
          type: string
          nullable: true
        parentId:
          type: string
          nullable: true
        teacherId:
          type: string
          nullable: true
        parentName:
          type: string
          nullable: true
        teacherName:
          type: string
          nullable: true
        attendanceRate:
          type: number
          nullable: true
        averageScore:
          type: number
          nullable: true
      required:
        - id
        - studentNo
        - firstName
        - lastName
        - grade
        - dateOfBirth
        - gender
        - enrollmentStatus
    CreateStudentRequest:
      type: object
      properties:
        firstName:
          type: string
          pattern: '^[a-zA-Z\s]+$'
        lastName:
          type: string
          pattern: '^[a-zA-Z\s]+$'
        grade:
          type: string
        dateOfBirth:
          type: string
          pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
        gender:
          type: string
          enum: [male, female]
        address:
          $ref: '#/components/schemas/Address'
          nullable: true
        faydaId:
          type: string
          pattern: '^[0-9]{12}$'
          nullable: true
        enrollmentStatus:
          type: string
          enum: [active, inactive, withdrawn, transferred, graduated]
        medicalInfo:
          type: string
          nullable: true
        emergencyContact:
          type: string
          nullable: true
        parentId:
          type: string
          nullable: true
        teacherId:
          type: string
          nullable: true
      required:
        - firstName
        - lastName
        - grade
        - dateOfBirth
    UpdateStudentRequest:
      type: object
      properties:
        firstName:
          type: string
          pattern: '^[a-zA-Z\s]+$'
        lastName:
          type: string
          pattern: '^[a-zA-Z\s]+$'
        grade:
          type: string
        dateOfBirth:
          type: string
          pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}$'
        gender:
          type: string
          enum: [male, female]
        address:
          $ref: '#/components/schemas/Address'
          nullable: true
        faydaId:
          type: string
          pattern: '^[0-9]{12}$'
          nullable: true
        enrollmentStatus:
          type: string
          enum: [active, inactive, withdrawn, transferred, graduated]
        medicalInfo:
          type: string
          nullable: true
        emergencyContact:
          type: string
          nullable: true
        parentId:
          type: string
          nullable: true
        teacherId:
          type: string
          nullable: true
...